tests moved from curve
